Fix location and update to new testing style
