+++ src/main/g8/$name__norm$/shared/src/main/scala/$name;format="lower,package,packaged"$/HelloWorld.scala
