--- src/main/g8/$name__norm$/src/test/scala/$name;format="lower,package,packaged"$/HelloWorldSpec.scala
@@ -1,26 +1,0 @@
-package $name;format="lower,package"$
-
-import zio._
-import zio.console._
-import zio.test._
-import zio.test.Assertion._
-import zio.test.environment._
-
-import HelloWorld._
-
-object HelloWorld {
-  def sayHello: ZIO[Console, Nothing, Unit] =
-    console.putStrLn("Hello, World!")
-}
-
-object HelloWorldSpec
-    extends DefaultRunnableSpec(
-      suite("HelloWorldSpec")(
-        testM("sayHello correctly displays output") {
-          for {
-            _      <- sayHello
-            output <- TestConsole.output
-          } yield assert(output, equalTo(Vector("Hello, World!\n")))
-        }
-      )
-    )
